Split the file I/O test helpers by consumer
`tests/common/file_io.rs` carried a module-wide `dead_code` expectation because
its twelve items fell into six different usage sets across the four scenario
binaries that included it. The expectation hid unused helpers from rustc, which is
what it was there to silence.

Fix the compilation boundary instead of suppressing the lint. Replace the file
with six modules, each included via `#[path]` only by the binaries that use all
of it:

- `durable_shell.rs` — all four binaries;
- `path_prompt.rs` — click-save-button, open and save (metadata asserts no
  prompts);
- `temp_svg.rs` — open, metadata and save;
- `temp_svg_write.rs` — open and metadata;
- `temp_svg_read.rs` — metadata and save;
- `temp_svg_exists.rs` — save only.

Two changes made that possible. `DurableShell::entity()` was the sole item not
used by all four binaries, so it is gone; the `entity` field is `pub(crate)` and
read directly, and since `with_visual_cx` reads it everywhere no field is dead
either. `TempSvgFile`'s `file_name` and `cleanup` fields are likewise
`pub(crate)`, so the three operation modules can reach them from sibling modules.

No module needs a `dead_code` expectation now, so an unused file I/O helper will
fail the lint gate again. The one remaining expectation is in `shared_helpers.rs`,
whose forty-one consumers would need roughly twenty modules to split the same way;
issue #150 tracks that, and its module documentation now points at this layout as
the pattern to follow.
