+++ tests/common/temp_svg.rs
@@ -1,0 +1,59 @@
+//! A scenario-owned temporary SVG file reached through a cap-std capability.
+//!
+//! The operations on [`TempSvgFile`] are split across sibling modules so each
+//! scenario binary compiles only the ones it drives: see `temp_svg_write.rs`,
+//! `temp_svg_read.rs` and `temp_svg_exists.rs`. The fields are `pub(crate)` so
+//! those modules can reach them. Everything defined here is used by every binary
+//! that includes it, so no `dead_code` expectation is needed.
+
+use camino::{Utf8Path, Utf8PathBuf};
+use cap_std::{ambient_authority, fs_utf8::Dir};
+use test_support::{TestSupportError, TestSupportResult};
+use uuid::Uuid;
+
+use crate::common::TempFileGuard;
+
+/// A uniquely named temporary SVG whose lifetime is owned by the scenario.
+pub struct TempSvgFile {
+    path: Utf8PathBuf,
+    /// Name within the owning directory capability, used by the operation modules.
+    pub(crate) file_name: Utf8PathBuf,
+    /// Directory capability and drop-time cleanup, used by the operation modules.
+    pub(crate) cleanup: TempFileGuard,
+}
+
+impl TempSvgFile {
+    /// Creates a uniquely named, UUID-suffixed SVG file path under the system temporary
+    /// directory and prepares it for cleanup on drop.
+    ///
+    /// The file itself is not created on disc until it is written; only the owning directory
+    /// capability and cleanup guard are established here.
+    ///
+    /// # Errors
+    ///
+    /// Returns `Err` if the system temporary directory is not valid UTF-8, or if it cannot be
+    /// opened as a cap-std directory capability.
+    pub fn create(prefix: &str) -> TestSupportResult<Self> {
+        let temp_dir = Utf8PathBuf::from_path_buf(std::env::temp_dir()).map_err(|path| {
+            TestSupportError::expectation(format!(
+                "temporary directory path is not valid UTF-8: {}",
+                path.display()
+            ))
+        })?;
+        let file_name = Utf8PathBuf::from(format!("{prefix}-{}.svg", Uuid::new_v4()));
+        let path = temp_dir.join(&file_name);
+        let dir = Dir::open_ambient_dir(&temp_dir, ambient_authority())
+            .map_err(|error| TestSupportError::io("opening the temporary directory", error))?;
+        let cleanup = TempFileGuard::new(dir, file_name.clone());
+        Ok(Self {
+            path,
+            file_name,
+            cleanup,
+        })
+    }
+
+    /// Returns the full path of the temporary file, whether or not it has been written yet.
+    pub fn path(&self) -> &Utf8Path {
+        self.path.as_path()
+    }
+}
